examples/lipsync-demo/requirements.txt
# Text-to-Speech dependencies
google-genai

# Speech-to-Text dependencies  
google-cloud-speech

# Additional utilities
python-dotenv